Hiển thị sản phẩm ra trang chủ

+++ NoiThat_v2.0/NoiThat_v2.0/Areas/Admin/Controllers/DanhMucController.cs
@@ -122,7 +122,9 @@ namespace NoiThat_v2._0.Areas.Admin.Controllers
                     DanhMucSanPham d = db.DanhMucSanPhams.Where(p => p.ID == id).FirstOrDefault();
                     NhomSanPham n = db.NhomSanPhams.Where(p => p.ID == d.IDNhomSP).FirstOrDefault();
                     string f = Server.MapPath(string.Format("~/storage/{0}/{1}", n.Ten_slug, d.Ten_slug));
-                    Directory.Delete(f);
+
+                    if(Directory.Exists(f))
+                        Directory.Delete(f);
 
                     db.DanhMucSanPhams.Remove(d);
                     db.SaveChanges();
